+++ src/core/download/encryption/mpeg_ts_sample_aes/h264.ts
@@ -1,4 +1,4 @@
-import { CbcDecryptJob, decryptCbcJobs } from "./aes_blocks";
+import { CbcDecryptJob, decryptCbcJobs } from "../sample_aes/aes_blocks";
 
 interface AnnexBNalUnit {
     readonly prefix: Buffer;